Customer Facts creation + Customers namespaces refactor
