+++ src/Resources/Programmatic/Programmatic.php
@@ -3,6 +3,7 @@
 namespace Bildvitta\IssCrm\Resources\Programmatic;
 
 use Bildvitta\IssCrm\IssCrm;
+use Bildvitta\IssCrm\Resources\Programmatic\Customers\Customers;
 
 class Programmatic
 {
